+++ src/main/kotlin/io/redgreen/timelapse/changedfiles/ChangedFilesModel.kt
@@ -11,25 +11,25 @@ import io.redgreen.timelapse.vcs.ChangedFile
 
 data class ChangedFilesModel(
   val commitId: String?,
-  val filePath: String?,
+  val selectedFilePath: String?,
   val getChangedFilesAsyncOp: AsyncOp<List<ChangedFile>, GetChangedFilesFailure> = idle(),
 ) {
   companion object {
     fun noFileAndRevisionSelected(): ChangedFilesModel =
-      ChangedFilesModel(commitId = null, filePath = null)
+      ChangedFilesModel(commitId = null, selectedFilePath = null)
   }
 
   fun fileAndRevisionSelected(
-    filePath: String,
+    selectedFilePath: String,
     commitId: String,
   ): ChangedFilesModel =
-    copy(commitId = commitId, filePath = filePath, getChangedFilesAsyncOp = inFlight())
+    copy(commitId = commitId, selectedFilePath = selectedFilePath, getChangedFilesAsyncOp = inFlight())
 
   fun noOtherFilesChanged(): ChangedFilesModel =
     copy(getChangedFilesAsyncOp = content(emptyList()))
 
   fun someMoreFilesChanged(changedFiles: List<ChangedFile>): ChangedFilesModel =
-    copy(getChangedFilesAsyncOp = content(changedFiles))
+    copy(getChangedFilesAsyncOp = content(changedFiles.filter { it.filePath != selectedFilePath }))
 
   fun gettingChangedFilesFailed(): ChangedFilesModel =
     copy(getChangedFilesAsyncOp = failure(Unknown))